+++ .github/workflows/ci.yml
@@ -135,8 +135,8 @@ jobs:
       - name: Check Import paths in Docs
         run: "deno task lint:doc-imports"
 
-  hash-wasm:
-    name: "_wasm_crypto/"
+  crypto-wasm:
+    name: "crypto/_wasm"
     runs-on: ${{ github.repository_owner == 'denoland' && 'ubuntu-20.04-xl' || 'ubuntu-20.04' }}
     steps:
       - name: Clone repository
@@ -147,12 +147,12 @@ jobs:
           submodules: false
           persist-credentials: false
 
-      - name: Check for changes to _wasm_crypto/
+      - name: Check for changes to crypto/_wasm
         id: source
         run: |-
           set -o errexit
           shopt -s inherit_errexit
-          declare modifications="$(git diff --name-only HEAD~ -- ./crypto/_wasm_crypto)"
+          declare modifications="$(git diff --name-only HEAD~ -- ./crypto/_wasm)"
           declare modified="$([[ "$modifications" ]] && echo true || echo false)"
           echo "modified=$modified" >> $GITHUB_OUTPUT
           echo "Wasm crypto source modified in this commit? $modified"
@@ -166,7 +166,7 @@ jobs:
         uses: hecrj/setup-rust-action@v1
         if: success() && steps.source.outputs.modified == 'true'
         with:
-          # This must match the version in crypto/_wasm_crypto/rust-toolchain.toml
+          # This must match the version in crypto/_wasm/rust-toolchain.toml
           rust-version: 1.64.0
           targets: wasm32-unknown-unknown
           components: rustfmt
