+++ _tools/check_deprecation.ts
@@ -12,7 +12,7 @@ const EXCLUDED_PATHS = [
   "dotenv/testdata",
   "fs/testdata",
   "http/testdata",
-  "crypto/_wasm_crypto/target",
+  "crypto/_wasm/target",
   "encoding/varint/_wasm_varint/target",
   "hash/_wasm/target",
 ];
