+++ _tools/check_licence.ts
@@ -10,7 +10,7 @@ const EXCLUDED_DIRS = [
   "fs/testdata",
   "http/testdata",
   "node/testdata",
-  "crypto/_wasm_crypto/target",
+  "crypto/_wasm/target",
   "encoding/varint/_wasm_varint/target",
   "_tools/testdata",
 ];
