+++ crypto/_wasm/src/digest.rs
