+++ deno.json
@@ -8,7 +8,7 @@
     "files": {
       "exclude": [
         ".git",
-        "crypto/_wasm_crypto/target",
+        "crypto/_wasm/target",
         "encoding/varint/_wasm_varint/target",
         "cov",
         "encoding/testdata/jsonc",
@@ -25,7 +25,7 @@
     "files": {
       "exclude": [
         ".git",
-        "crypto/_wasm_crypto/target",
+        "crypto/_wasm/target",
         "encoding/varint/_wasm_varint/target",
         "encoding/testdata/jsonc",
         "encoding/front_matter/testdata",
@@ -52,7 +52,7 @@
     "node:list-remaining-tests": "deno run --allow-net --allow-write --allow-read node/_tools/list_remaining_tests.ts",
     "node:setup": "deno task node:list-remaining-tests && deno run --allow-read --allow-net --allow-write node/_tools/setup.ts",
     "node:check-unstable-api": "deno check --unstable ./node/module_all.ts",
-    "build:crypto": "deno task --cwd crypto/_wasm_crypto wasmbuild",
+    "build:crypto": "deno task --cwd crypto/_wasm wasmbuild",
     "build:varint": "deno task --cwd encoding/varint/_wasm_varint wasmbuild",
     "wasmbuild": "deno run --unstable -A https://deno.land/x/wasmbuild@0.10.0/main.ts --js-ext mjs --sync"
   }
