+++ node/internal/crypto/util.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 // Copyright Joyent, Inc. and Node.js contributors. All rights reserved. MIT license.
 
-import { digestAlgorithms } from "../../../crypto/_wasm_crypto/mod.ts";
+import { digestAlgorithms } from "../../../crypto/_wasm/mod.ts";
 import { getCiphers } from "../../_crypto/crypto_browserify/browserify_aes/mod.js";
 import { notImplemented } from "../../_utils.ts";
 import { Buffer } from "../../buffer.ts";
